Unit test for the app (#59)
* Adding an app test

* fixup: missed rebase message

* fix file paths

* fixup: escaped percent sign
