+++ Dockerfile
@@ -19,4 +19,4 @@ EXPOSE 8501
 
 HEALTHCHECK CMD curl --fail http://localhost:8501/_stcore/health
 
-ENTRYPOINT ["poetry", "run", "streamlit", "run", "scripts/widget.py", "--server.port=8501", "--server.address=0.0.0.0"]
+ENTRYPOINT ["poetry", "run", "streamlit", "run", "ngm/widget.py", "--server.port=8501", "--server.address=0.0.0.0"]
